renamed the colorgen files to more appropriate nvtx based name
